Split off templates from resolver
