+++ contracts/warp-controller/examples/warp-controller-schema.rs
@@ -4,7 +4,7 @@ use std::fs::create_dir_all;
 use controller::{
     account::{AccountResponse, AccountsResponse},
     job::{JobResponse, JobsResponse},
-    QueryMsg, {Config, ConfigResponse, ExecuteMsg, InstantiateMsg, SimulateResponse},
+    QueryMsg, {Config, ConfigResponse, ExecuteMsg, InstantiateMsg},
 };
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 use cosmwasm_std::{CosmosMsg, QueryRequest};
@@ -24,7 +24,6 @@ fn main() {
     export_schema(&schema_for!(JobsResponse), &out_dir);
     export_schema(&schema_for!(AccountResponse), &out_dir);
     export_schema(&schema_for!(AccountsResponse), &out_dir);
-    export_schema(&schema_for!(SimulateResponse), &out_dir);
     export_schema(&schema_for!(CosmosMsg), &out_dir);
     export_schema(&schema_for!(QueryRequest<String>), &out_dir);
 }
