+++ contracts/warp-controller/src/query/controller.rs
@@ -1,6 +1,6 @@
 use crate::state::CONFIG;
 use controller::{ConfigResponse, QueryConfigMsg};
-use cosmwasm_std::{Deps, Env, StdError, StdResult};
+use cosmwasm_std::{Deps, Env, StdResult};
 
 pub fn query_config(deps: Deps, _env: Env, _data: QueryConfigMsg) -> StdResult<ConfigResponse> {
     let config = CONFIG.load(deps.storage)?;
