+++ contracts/warp-templates/Cargo.toml
@@ -1,0 +1,50 @@
+[package]
+name = "warp-templates"
+version = "0.1.0"
+authors = ["Terra Money <core@terra.money>"]
+edition = "2021"
+
+exclude = [
+  # Those files are rust-optimizer artifacts. You might want to commit them for convenience but they should not be part of the source code publication.
+  "contract.wasm",
+  "hash.txt",
+]
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[features]
+# for more explicit tests, cargo test --features=backtraces
+backtraces = ["cosmwasm-std/backtraces"]
+# use library feature to disable all instantiate/execute/query exports
+library = []
+
+[package.metadata.scripts]
+optimize = """docker run --rm -v "${process.cwd()}":/code \
+  -v "${path.join(process.cwd(), "../../", "packages")}":/packages \
+  --mount type=volume,source="${contract}_cache",target=/code/target \
+  --mount type=volume,source=registry_cache,target=/usr/local/cargo/registry \
+  cosmwasm/rust-optimizer${process.env.TERRARIUM_ARCH_ARM64 ? "-arm64" : ""}:0.12.6
+"""
+
+[dependencies]
+cosmwasm-std = "1.1"
+cosmwasm-storage = "1.1"
+cosmwasm-schema = "1.1"
+base64 = "0.13.0"
+cw-asset = "2.2"
+cw-storage-plus = "0.16"
+cw2 = "0.16"
+cw20 = "0.16"
+cw721 = "0.16.0"
+templates = { path = "../../packages/templates",  default-features = false, version = "*" }
+resolver = { path = "../../packages/resolver", default-features = false, version = "*" }
+schemars = "0.8"
+thiserror = "1"
+serde-json-wasm = "0.4.1"
+json-codec-wasm = "0.1.0"
+
+[dev-dependencies]
+cw-multi-test = "0.16.0"
