+++ contracts/warp-templates/src/error.rs
