+++ contracts/warp-templates/src/state.rs
@@ -1,6 +1,6 @@
 use cw_storage_plus::{Item, Map};
-use resolver::{Config, State};
-use resolver::template::Template;
+use templates::template::Template;
+use templates::{Config, State};
 
 pub const CONFIG: Item<Config> = Item::new("config");
 pub const TEMPLATES: Map<u64, Template> = Map::new("templates");
