+++ contracts/warp-templates/src/tests.rs
@@ -1,0 +1,1 @@
+
