+++ packages/templates/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "controller"
+name = "templates"
 version = "0.1.0"
 authors = ["Terra Money <core@terra.money>"]
 edition = "2021"
@@ -18,9 +18,12 @@ cw-storage-plus = "0.16"
 cw2 = "0.16"
 schemars = "0.8"
 serde = { version = "1", default-features = false, features = ["derive"] }
+resolver = { path = "../resolver", default-features = false, version = "*" }
+json-codec-wasm = "0.1.0"
 strum = "0.24"
 strum_macros = "0.24"
 thiserror = { version = "1" }
+controller = {path = "../controller"}
 
 [dev-dependencies]
 cw-multi-test = "0.16"
