+++ packages/templates/examples/account-schema.rs
@@ -1,9 +1,9 @@
 use std::env::current_dir;
 use std::fs::create_dir_all;
 
+use controller::QueryMsg;
+use controller::{ExecuteMsg, InstantiateMsg};
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
-use resolver::{ExecuteMsg, InstantiateMsg, QueryMsg
-};
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
